fix: prevent images from appearing multiple times in generated readme and changelog

+++ scripts/generate-included-readme.js
@@ -241,6 +241,13 @@ fs.readFile(changeLogPath, (err, data) => {
     console.error(err)
   }
 
+  const imageExists = data.includes(versionTag)
+
+  if (imageExists) {
+    console.log(`Image already exists in README and CHANGELOG.`)
+    process.exit(1)
+  }
+
   const updatedChangeLog = `${data.toString()} \n[cypress/included:${versionTag}](${versionTag}) | \`${baseImageTag}\``
 
   fs.writeFileSync(changeLogPath, updatedChangeLog.trim() + "\n", "utf8")
